Created a View Package and moved view.java there
