+++ src/main/java/earthGame/Model/Card.java
@@ -2,7 +2,7 @@ package earthGame.Model;
 
 import java.util.ArrayList;
 
-class Card {
+public class Card {
     public String name;
     public String subName;
     public ArrayList<String> habitatElements = new ArrayList<String>();
